Move is_standby_cluster Remove unnecessary scenario Renaming and styling improvements
