+++ features/standby_cluster.feature
@@ -11,10 +11,6 @@ Feature: standby cluster
     Then postgres2 role is the replica after 24 seconds
     And table foo is present on postgres2 after 20 seconds
 
-  Scenario: check failover with dead postmaster
-    When I kill postmaster on postgres1
-	Then postgres2 is replicating from postgres1 after 20 seconds
-
   Scenario: check failover
     When I kill postgres1
 	Then postgres2 is replicating from postgres0 after 20 seconds
